move password handling functions from user.controller.js to external file (helpers/passwordHandling.js)
